Move && Improve InvestmentPreview, Button, SectionHeader components
