+++ app/components/dashboard/Dashboard.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
 import { Col, Row } from "reactstrap";
 import { LayoutAuthorized } from "../layouts/LayoutAuthorized";
+import { InvestmentPreview } from "../shared/InvestmentPreview";
 import { SectionHeader } from "../shared/SectionHeader";
 import { Tag } from "../shared/Tag";
-import { InvestmentPreview } from "./investmentOportunities/InvestmentPreview";
 import { MyPortfolioWidget } from "./myPortfolio/MyPortfolioWidget";
 import { MyWalletWidget } from "./myWallet/MyWalletWidget";
 import { NeufundKpiWidget } from "./NeufundKpiWidget";
@@ -58,7 +58,7 @@ export const Dashboard = () => (
       </Col>
     </Row>
     <SectionHeader>
-      investment opportunities{" "}
+      investment opportunities
       <Tag className="ml-3" theme="green" layout="ghost" size="small" text="1 new" />
     </SectionHeader>
     <Row className="py-4">
@@ -70,7 +70,7 @@ export const Dashboard = () => (
           tokenPrice={"2€"}
           neuInvestorsNum={500}
           startingOn="22.12.2019"
-          handleEmailSend={() => {}}
+          handleEmailSend={() => { }}
           endInDays={25}
           company={"Superawesome startup 2"}
           hasStarted={true}
@@ -99,7 +99,7 @@ export const Dashboard = () => (
           tokenPrice={"2€"}
           neuInvestorsNum={500}
           startingOn="22.12.2019"
-          handleEmailSend={() => {}}
+          handleEmailSend={() => { }}
           endInDays={25}
           company={"Superawesome startup one"}
           hasStarted={false}
