+++ app/components/shared/InvestmentPreview.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/neufund-theme";
+@import "../../styles/neufund-theme";
 
 .investment-preview {
   background-color: $neu-white-two;
