+++ app/components/shared/SectionHeader.tsx
@@ -1,6 +1,12 @@
+import * as cn from "classnames";
 import * as React from "react";
+
 import * as styles from "./SectionHeader.module.scss";
 
-export const SectionHeader: React.SFC = ({ children }) => (
-  <h3 className={styles.sectionHeader}>{children}</h3>
+interface IProps {
+  className?: string;
+}
+
+export const SectionHeader: React.SFC<IProps> = ({ children, className }) => (
+  <h3 className={cn(styles.sectionHeader, className)}>{children}</h3>
 );
